Refactor game canvas to use hooks and extract utilities
This commit refactors the game canvas component by extracting logic into custom hooks and utility functions. This improves code organization and maintainability.

Key changes include:
- **Extracted Hooks:** `useCars`, `useCrimeIncidents`, `usePedestrians`, and `useEmergencyVehicles` hooks are introduced to manage the state and logic for cars, crime incidents, pedestrians, and emergency vehicles respectively.
- **Extracted Drawing Functions:** `drawCars` and `drawEmergencyVehicles` functions are moved to separate files for better modularity.
- **Extracted Utility Functions:** Helper functions for building and viewport manipulation are moved to their respective utility modules (`buildingHelpers.ts` and `viewportHelpers.ts`).

These changes aim to simplify the main `CanvasIsometricGrid` component and make the game logic more testable and reusable.
