+++ src/components/game/CanvasIsometricGrid.tsx
@@ -119,6 +119,10 @@ import {
 } from '@/components/game/renderHelpers';
 import { drawAirplanes as drawAirplanesUtil, drawHelicopters as drawHelicoptersUtil } from '@/components/game/drawAircraft';
 import { drawPedestrians as drawPedestriansUtil } from '@/components/game/drawPedestrians';
+import { useCars, useCrimeIncidents, usePedestrians, useEmergencyVehicles } from '@/components/game/hooks';
+import { drawCars as drawCarsUtil, drawEmergencyVehicles as drawEmergencyVehiclesUtil } from '@/components/game/drawing';
+import { isPartOfMultiTileBuilding, findBuildingOrigin, isPartOfParkBuilding } from '@/components/game/utils/buildingHelpers';
+import { clampOffset, getTouchDistance, getTouchCenter, getMapBounds as getMapBoundsUtil } from '@/components/game/utils/viewportHelpers';
 
 // Props interface for CanvasIsometricGrid
 export interface CanvasIsometricGridProps {
@@ -153,19 +157,38 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
     screenY: number;
   } | null>(null);
   const [zoom, setZoom] = useState(isMobile ? 0.6 : 1);
-  const carsRef = useRef<Car[]>([]);
-  const carIdRef = useRef(0);
-  const carSpawnTimerRef = useRef(0);
-  const emergencyVehiclesRef = useRef<EmergencyVehicle[]>([]);
-  const emergencyVehicleIdRef = useRef(0);
-  const emergencyDispatchTimerRef = useRef(0);
+  
+  // Track fires and crimes for emergency vehicle dispatch
   const activeFiresRef = useRef<Set<string>>(new Set()); // Track fires that already have a truck dispatched
   const activeCrimesRef = useRef<Set<string>>(new Set()); // Track crimes that already have a car dispatched
-  const activeCrimeIncidentsRef = useRef<Map<string, { x: number; y: number; type: 'robbery' | 'burglary' | 'disturbance' | 'traffic'; timeRemaining: number }>>(new Map()); // Persistent crime incidents
-  const crimeSpawnTimerRef = useRef(0); // Timer for spawning new crime incidents
   
-  // Pedestrian system refs
-  const pedestriansRef = useRef<Pedestrian[]>([]);
+  // Performance: Cache expensive grid calculations
+  const cachedRoadTileCountRef = useRef<{ count: number; gridVersion: number }>({ count: 0, gridVersion: -1 });
+  const cachedPopulationRef = useRef<{ count: number; gridVersion: number }>({ count: 0, gridVersion: -1 });
+  const gridVersionRef = useRef(0);
+  
+  // Use extracted hooks for entity management
+  const { carsRef, updateCars } = useCars(worldStateRef, isMobile);
+  const { activeCrimeIncidentsRef, spawnCrimeIncidents, updateCrimeIncidents, findCrimeIncidents } = useCrimeIncidents(
+    worldStateRef,
+    state.services.police,
+    state.stats.population
+  );
+  const { pedestriansRef, updatePedestrians } = usePedestrians(
+    worldStateRef,
+    gridVersionRef,
+    cachedRoadTileCountRef,
+    isMobile
+  );
+  const { emergencyVehiclesRef, updateEmergencyVehicles } = useEmergencyVehicles(
+    worldStateRef,
+    activeFiresRef,
+    activeCrimesRef,
+    activeCrimeIncidentsRef,
+    findCrimeIncidents
+  );
+  
+  // Pedestrian system refs (kept for compatibility - now managed by hook)
   const pedestrianIdRef = useRef(0);
   const pedestrianSpawnTimerRef = useRef(0);
   
@@ -204,11 +227,6 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
   // Factory smog system refs
   const factorySmogRef = useRef<FactorySmog[]>([]);
   const smogLastGridVersionRef = useRef(-1); // Track when to rebuild factory list
-
-  // Performance: Cache expensive grid calculations
-  const cachedRoadTileCountRef = useRef<{ count: number; gridVersion: number }>({ count: 0, gridVersion: -1 });
-  const cachedPopulationRef = useRef<{ count: number; gridVersion: number }>({ count: 0, gridVersion: -1 });
-  const gridVersionRef = useRef(0);
 
   const worldStateRef = useRef<WorldRenderState>({
     grid,
@@ -335,44 +353,17 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
     };
   }, []);
 
-  const spawnRandomCar = useCallback(() => {
-    const { grid: currentGrid, gridSize: currentGridSize } = worldStateRef.current;
-    if (!currentGrid || currentGridSize <= 0) return false;
-    
-    for (let attempt = 0; attempt < 20; attempt++) {
-      const tileX = Math.floor(Math.random() * currentGridSize);
-      const tileY = Math.floor(Math.random() * currentGridSize);
-      if (!isRoadTile(currentGrid, currentGridSize, tileX, tileY)) continue;
-      
-      const options = getDirectionOptions(currentGrid, currentGridSize, tileX, tileY);
-      if (options.length === 0) continue;
-      
-      const direction = options[Math.floor(Math.random() * options.length)];
-      carsRef.current.push({
-        id: carIdRef.current++,
-        tileX,
-        tileY,
-        direction,
-        progress: Math.random() * 0.8,
-        speed: (0.35 + Math.random() * 0.35) * 0.7,
-        age: 0,
-        maxAge: 1800 + Math.random() * 2700,
-        color: CAR_COLORS[Math.floor(Math.random() * CAR_COLORS.length)],
-        laneOffset: (Math.random() < 0.5 ? -1 : 1) * (4 + Math.random() * 3),
-      });
-      return true;
-    }
-    
-    return false;
-  }, []);
+  // Car spawning and updating now handled by useCars hook
 
   // Find residential buildings for pedestrian spawning (uses imported utility)
+  // Memoized to avoid recreating on every render
   const findResidentialBuildingsCallback = useCallback((): { x: number; y: number }[] => {
     const { grid: currentGrid, gridSize: currentGridSize } = worldStateRef.current;
     return findResidentialBuildings(currentGrid, currentGridSize);
   }, []);
 
   // Find destinations for pedestrians (uses imported utility)
+  // Memoized to avoid recreating on every render
   const findPedestrianDestinationsCallback = useCallback((): { x: number; y: number; type: PedestrianDestType }[] => {
     const { grid: currentGrid, gridSize: currentGridSize } = worldStateRef.current;
     return findPedestrianDestinations(currentGrid, currentGridSize);
@@ -460,783 +451,17 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
     return findFires(currentGrid, currentGridSize);
   }, []);
 
-  // Spawn new crime incidents periodically (persistent like fires)
-  const spawnCrimeIncidents = useCallback((delta: number) => {
-    const { grid: currentGrid, gridSize: currentGridSize, speed: currentSpeed } = worldStateRef.current;
-    if (!currentGrid || currentGridSize <= 0 || currentSpeed === 0) return;
-    
-    const speedMultiplier = currentSpeed === 1 ? 1 : currentSpeed === 2 ? 2 : 3;
-    crimeSpawnTimerRef.current -= delta * speedMultiplier;
-    
-    // Spawn new crimes every 3-5 seconds (game time adjusted)
-    if (crimeSpawnTimerRef.current > 0) return;
-    crimeSpawnTimerRef.current = 3 + Math.random() * 2;
-    
-    // Collect eligible tiles for crime (buildings with activity)
-    const eligibleTiles: { x: number; y: number; policeCoverage: number }[] = [];
-    
-    for (let y = 0; y < currentGridSize; y++) {
-      for (let x = 0; x < currentGridSize; x++) {
-        const tile = currentGrid[y][x];
-        // Only consider populated buildings (residential/commercial/industrial)
-        // FIX: Proper parentheses for operator precedence
-        const isBuilding = tile.building.type !== 'grass' && 
-            tile.building.type !== 'water' && 
-            tile.building.type !== 'road' && 
-            tile.building.type !== 'tree' &&
-            tile.building.type !== 'empty';
-        const hasActivity = tile.building.population > 0 || tile.building.jobs > 0;
-        
-        if (isBuilding && hasActivity) {
-          const policeCoverage = state.services.police[y]?.[x] || 0;
-          // Crime can happen anywhere, but more likely in low-coverage areas
-          eligibleTiles.push({ x, y, policeCoverage });
-        }
-      }
-    }
-    
-    if (eligibleTiles.length === 0) return;
-    
-    // Determine how many new crimes to spawn (based on city size and coverage)
-    const avgCoverage = eligibleTiles.reduce((sum, t) => sum + t.policeCoverage, 0) / eligibleTiles.length;
-    const baseChance = avgCoverage < 20 ? 0.4 : avgCoverage < 40 ? 0.25 : avgCoverage < 60 ? 0.15 : 0.08;
-    
-    // Max active crimes based on population (more people = more potential crime)
-    const population = state.stats.population;
-    const maxActiveCrimes = Math.max(2, Math.floor(population / 500));
-    
-    if (activeCrimeIncidentsRef.current.size >= maxActiveCrimes) return;
-    
-    // Try to spawn 1-2 crimes
-    const crimesToSpawn = Math.random() < 0.3 ? 2 : 1;
-    
-    for (let i = 0; i < crimesToSpawn; i++) {
-      if (activeCrimeIncidentsRef.current.size >= maxActiveCrimes) break;
-      if (Math.random() > baseChance) continue;
-      
-      // Weight selection toward low-coverage areas
-      const weightedTiles = eligibleTiles.filter(t => {
-        const key = `${t.x},${t.y}`;
-        if (activeCrimeIncidentsRef.current.has(key)) return false;
-        // Higher weight for lower coverage
-        const weight = Math.max(0.1, 1 - t.policeCoverage / 100);
-        return Math.random() < weight;
-      });
-      
-      if (weightedTiles.length === 0) continue;
-      
-      const target = weightedTiles[Math.floor(Math.random() * weightedTiles.length)];
-      const key = `${target.x},${target.y}`;
-      
-      // Different crime types with different durations
-      const crimeTypes: Array<'robbery' | 'burglary' | 'disturbance' | 'traffic'> = ['robbery', 'burglary', 'disturbance', 'traffic'];
-      const crimeType = crimeTypes[Math.floor(Math.random() * crimeTypes.length)];
-      const duration = crimeType === 'traffic' ? 15 : crimeType === 'disturbance' ? 20 : 30; // Seconds to resolve if no police
-      
-      activeCrimeIncidentsRef.current.set(key, {
-        x: target.x,
-        y: target.y,
-        type: crimeType,
-        timeRemaining: duration,
-      });
-    }
-  }, [state.services.police, state.stats.population]);
-  
-  // Update crime incidents (decay over time if not responded to)
-  const updateCrimeIncidents = useCallback((delta: number) => {
-    const { speed: currentSpeed } = worldStateRef.current;
-    if (currentSpeed === 0) return;
-    
-    const speedMultiplier = currentSpeed === 1 ? 1 : currentSpeed === 2 ? 2 : 3;
-    const keysToDelete: string[] = [];
-    
-    // Iterate and track which crimes to delete
-    activeCrimeIncidentsRef.current.forEach((crime, key) => {
-      // If police car is responding, don't decay
-      if (activeCrimesRef.current.has(key)) return;
-      
-      // Update time remaining by creating a new crime object
-      const newTimeRemaining = crime.timeRemaining - delta * speedMultiplier;
-      if (newTimeRemaining <= 0) {
-        // Crime "resolved" without police (criminal escaped, situation de-escalated)
-        keysToDelete.push(key);
-      } else {
-        // Update the crime's time remaining
-        activeCrimeIncidentsRef.current.set(key, { ...crime, timeRemaining: newTimeRemaining });
-      }
-    });
-    
-    // Delete expired crimes
-    keysToDelete.forEach(key => activeCrimeIncidentsRef.current.delete(key));
-  }, []);
-  
-  // Find active crime incidents that need police response
-  const findCrimeIncidents = useCallback((): { x: number; y: number }[] => {
-    return Array.from(activeCrimeIncidentsRef.current.values()).map(c => ({ x: c.x, y: c.y }));
-  }, []);
+  // Crime incident spawning and updating now handled by useCrimeIncidents hook
 
-  // Dispatch emergency vehicle
-  const dispatchEmergencyVehicle = useCallback((
-    type: EmergencyVehicleType,
-    stationX: number,
-    stationY: number,
-    targetX: number,
-    targetY: number
-  ): boolean => {
-    const { grid: currentGrid, gridSize: currentGridSize } = worldStateRef.current;
-    if (!currentGrid || currentGridSize <= 0) return false;
+  // Emergency vehicle dispatch and updating now handled by useEmergencyVehicles hook
 
-    const path = findPathOnRoads(currentGrid, currentGridSize, stationX, stationY, targetX, targetY);
-    if (!path || path.length === 0) return false;
+  // Car updating now handled by useCars hook
 
-    const startTile = path[0];
-    let direction: CarDirection = 'south'; // Default direction
-    
-    // If path has at least 2 tiles, get direction from first to second
-    if (path.length >= 2) {
-      const nextTile = path[1];
-      const dir = getDirectionToTile(startTile.x, startTile.y, nextTile.x, nextTile.y);
-      if (dir) direction = dir;
-    }
-
-    emergencyVehiclesRef.current.push({
-      id: emergencyVehicleIdRef.current++,
-      type,
-      tileX: startTile.x,
-      tileY: startTile.y,
-      direction,
-      progress: 0,
-      speed: type === 'fire_truck' ? 0.8 : 0.9, // Emergency vehicles are faster
-      state: 'dispatching',
-      stationX,
-      stationY,
-      targetX,
-      targetY,
-      path,
-      pathIndex: 0,
-      respondTime: 0,
-      laneOffset: 0, // Emergency vehicles drive in the center
-      flashTimer: 0,
-    });
-
-    return true;
-  }, []);
-
-  // Update emergency vehicles dispatch logic
-  const updateEmergencyDispatch = useCallback(() => {
-    const { grid: currentGrid, gridSize: currentGridSize, speed: currentSpeed } = worldStateRef.current;
-    if (!currentGrid || currentGridSize <= 0 || currentSpeed === 0) return;
-    
-    const fires = findFiresCallback();
-    const fireStations = findStationsCallback('fire_station');
-    
-    for (const fire of fires) {
-      const fireKey = `${fire.x},${fire.y}`;
-      if (activeFiresRef.current.has(fireKey)) continue;
-      
-      // Find nearest fire station
-      let nearestStation: { x: number; y: number } | null = null;
-      let nearestDist = Infinity;
-      
-      for (const station of fireStations) {
-        const dist = Math.abs(station.x - fire.x) + Math.abs(station.y - fire.y);
-        if (dist < nearestDist) {
-          nearestDist = dist;
-          nearestStation = station;
-        }
-      }
-      
-      if (nearestStation) {
-        if (dispatchEmergencyVehicle('fire_truck', nearestStation.x, nearestStation.y, fire.x, fire.y)) {
-          activeFiresRef.current.add(fireKey);
-        }
-      }
-    }
-
-    // Find crimes that need police dispatched
-    const crimes = findCrimeIncidents();
-    const policeStations = findStationsCallback('police_station');
-    
-    // Limit police dispatches per update (increased for more action)
-    let dispatched = 0;
-    const maxDispatchPerCheck = Math.max(3, Math.min(6, policeStations.length * 2)); // Scale with stations
-    for (const crime of crimes) {
-      if (dispatched >= maxDispatchPerCheck) break;
-      
-      const crimeKey = `${crime.x},${crime.y}`;
-      if (activeCrimesRef.current.has(crimeKey)) continue;
-      
-      // Find nearest police station
-      let nearestStation: { x: number; y: number } | null = null;
-      let nearestDist = Infinity;
-      
-      for (const station of policeStations) {
-        const dist = Math.abs(station.x - crime.x) + Math.abs(station.y - crime.y);
-        if (dist < nearestDist) {
-          nearestDist = dist;
-          nearestStation = station;
-        }
-      }
-      
-      if (nearestStation) {
-        if (dispatchEmergencyVehicle('police_car', nearestStation.x, nearestStation.y, crime.x, crime.y)) {
-          activeCrimesRef.current.add(crimeKey);
-          dispatched++;
-        }
-      }
-    }
-  }, [findFiresCallback, findCrimeIncidents, findStationsCallback, dispatchEmergencyVehicle]);
-
-  // Update emergency vehicles movement and state
-  const updateEmergencyVehicles = useCallback((delta: number) => {
-    const { grid: currentGrid, gridSize: currentGridSize, speed: currentSpeed } = worldStateRef.current;
-    if (!currentGrid || currentGridSize <= 0) {
-      emergencyVehiclesRef.current = [];
-      return;
-    }
-
-    const speedMultiplier = currentSpeed === 0 ? 0 : currentSpeed === 1 ? 1 : currentSpeed === 2 ? 2.5 : 4;
-    
-    // Dispatch check every second or so
-    emergencyDispatchTimerRef.current -= delta;
-    if (emergencyDispatchTimerRef.current <= 0) {
-      updateEmergencyDispatch();
-      emergencyDispatchTimerRef.current = 1.5;
-    }
-
-    const updatedVehicles: EmergencyVehicle[] = [];
-    
-    for (const vehicle of [...emergencyVehiclesRef.current]) {
-      // Update flash timer for lights
-      vehicle.flashTimer += delta * 8;
-      
-      if (vehicle.state === 'responding') {
-        // Check if vehicle is still on a valid road (road might have been bulldozed)
-        if (!isRoadTile(currentGrid, currentGridSize, vehicle.tileX, vehicle.tileY)) {
-          const targetKey = `${vehicle.targetX},${vehicle.targetY}`;
-          if (vehicle.type === 'fire_truck') {
-            activeFiresRef.current.delete(targetKey);
-          } else {
-            activeCrimesRef.current.delete(targetKey);
-            activeCrimeIncidentsRef.current.delete(targetKey); // Also clear the crime incident
-          }
-          continue; // Remove vehicle
-        }
-        
-        // At the scene - spend some time responding
-        vehicle.respondTime += delta * speedMultiplier;
-        const respondDuration = vehicle.type === 'fire_truck' ? 8 : 5; // Fire trucks stay longer
-        
-        if (vehicle.respondTime >= respondDuration) {
-          // Done responding - crime is resolved, calculate return path
-          const targetKey = `${vehicle.targetX},${vehicle.targetY}`;
-          
-          // Clear the crime incident when police finish responding
-          if (vehicle.type === 'police_car') {
-            activeCrimeIncidentsRef.current.delete(targetKey);
-          }
-          
-          const returnPath = findPathOnRoads(
-            currentGrid, currentGridSize,
-            vehicle.tileX, vehicle.tileY,
-            vehicle.stationX, vehicle.stationY
-          );
-          
-          if (returnPath && returnPath.length >= 2) {
-            vehicle.path = returnPath;
-            vehicle.pathIndex = 0;
-            vehicle.state = 'returning';
-            vehicle.progress = 0;
-            
-            const nextTile = returnPath[1];
-            const dir = getDirectionToTile(vehicle.tileX, vehicle.tileY, nextTile.x, nextTile.y);
-            if (dir) vehicle.direction = dir;
-          } else if (returnPath && returnPath.length === 1) {
-            // Already at station's road - remove vehicle
-            if (vehicle.type === 'fire_truck') {
-              activeFiresRef.current.delete(targetKey);
-            } else {
-              activeCrimesRef.current.delete(targetKey);
-            }
-            continue;
-          } else {
-            // Can't find return path - remove vehicle and clear tracking
-            if (vehicle.type === 'fire_truck') {
-              activeFiresRef.current.delete(targetKey);
-            } else {
-              activeCrimesRef.current.delete(targetKey);
-            }
-            continue;
-          }
-        }
-        
-        updatedVehicles.push(vehicle);
-        continue;
-      }
-      
-      // Check if vehicle is still on a valid road
-      if (!isRoadTile(currentGrid, currentGridSize, vehicle.tileX, vehicle.tileY)) {
-        const targetKey = `${vehicle.targetX},${vehicle.targetY}`;
-        if (vehicle.type === 'fire_truck') {
-          activeFiresRef.current.delete(targetKey);
-        } else {
-          activeCrimesRef.current.delete(targetKey);
-          activeCrimeIncidentsRef.current.delete(targetKey); // Also clear the crime incident
-        }
-        continue;
-      }
-      
-      // Bounds check - remove vehicle if out of bounds
-      if (vehicle.tileX < 0 || vehicle.tileX >= currentGridSize || 
-          vehicle.tileY < 0 || vehicle.tileY >= currentGridSize) {
-        const targetKey = `${vehicle.targetX},${vehicle.targetY}`;
-        if (vehicle.type === 'fire_truck') {
-          activeFiresRef.current.delete(targetKey);
-        } else {
-          activeCrimesRef.current.delete(targetKey);
-          activeCrimeIncidentsRef.current.delete(targetKey); // Also clear the crime incident
-        }
-        continue; // Remove vehicle
-      }
-      
-      // Move vehicle along path
-      vehicle.progress += vehicle.speed * delta * speedMultiplier;
-      
-      let shouldRemove = false;
-      
-      // Handle edge case: path has only 1 tile (already at destination)
-      if (vehicle.path.length === 1 && vehicle.state === 'dispatching') {
-        vehicle.state = 'responding';
-        vehicle.respondTime = 0;
-        vehicle.progress = 0;
-        updatedVehicles.push(vehicle);
-        continue;
-      }
-      
-      while (vehicle.progress >= 1 && vehicle.pathIndex < vehicle.path.length - 1) {
-        vehicle.pathIndex++;
-        vehicle.progress -= 1;
-        
-        const currentTile = vehicle.path[vehicle.pathIndex];
-        
-        // Validate the next tile is in bounds
-        if (currentTile.x < 0 || currentTile.x >= currentGridSize || 
-            currentTile.y < 0 || currentTile.y >= currentGridSize) {
-          shouldRemove = true;
-          break;
-        }
-        
-        vehicle.tileX = currentTile.x;
-        vehicle.tileY = currentTile.y;
-        
-        // Check if reached destination
-        if (vehicle.pathIndex >= vehicle.path.length - 1) {
-          if (vehicle.state === 'dispatching') {
-            // Arrived at emergency scene
-            vehicle.state = 'responding';
-            vehicle.respondTime = 0;
-            vehicle.progress = 0; // Reset progress to keep vehicle centered on road tile
-          } else if (vehicle.state === 'returning') {
-            // Arrived back at station - remove vehicle
-            shouldRemove = true;
-          }
-          break;
-        }
-        
-        // Update direction for next segment
-        if (vehicle.pathIndex + 1 < vehicle.path.length) {
-          const nextTile = vehicle.path[vehicle.pathIndex + 1];
-          const dir = getDirectionToTile(vehicle.tileX, vehicle.tileY, nextTile.x, nextTile.y);
-          if (dir) vehicle.direction = dir;
-        }
-      }
-      
-      if (shouldRemove) {
-        const targetKey = `${vehicle.targetX},${vehicle.targetY}`;
-        if (vehicle.type === 'fire_truck') {
-          activeFiresRef.current.delete(targetKey);
-        } else {
-          activeCrimesRef.current.delete(targetKey);
-          activeCrimeIncidentsRef.current.delete(targetKey); // Also clear the crime incident
-        }
-        continue; // Don't add to updated list
-      }
-      
-      updatedVehicles.push(vehicle);
-    }
-    
-    emergencyVehiclesRef.current = updatedVehicles;
-  }, [updateEmergencyDispatch]);
-
-  const updateCars = useCallback((delta: number) => {
-    const { grid: currentGrid, gridSize: currentGridSize, speed: currentSpeed } = worldStateRef.current;
-    if (!currentGrid || currentGridSize <= 0) {
-      carsRef.current = [];
-      return;
-    }
-    
-    // Speed multiplier: 0 = paused, 1 = normal, 2 = fast (2x), 3 = very fast (4x)
-    const speedMultiplier = currentSpeed === 0 ? 0 : currentSpeed === 1 ? 1 : currentSpeed === 2 ? 2.5 : 4;
-    
-    // Reduce max cars on mobile for better performance
-    const baseMaxCars = 160;
-    const maxCars = Math.min(baseMaxCars, Math.max(16, Math.floor(currentGridSize * (2))));
-    carSpawnTimerRef.current -= delta;
-    if (carsRef.current.length < maxCars && carSpawnTimerRef.current <= 0) {
-      if (spawnRandomCar()) {
-        carSpawnTimerRef.current = 0.9 + Math.random() * 1.3;
-      } else {
-        carSpawnTimerRef.current = 0.5;
-      }
-    }
-    
-    const updatedCars: Car[] = [];
-    for (const car of [...carsRef.current]) {
-      let alive = true;
-      
-      car.age += delta;
-      if (car.age > car.maxAge) {
-        continue;
-      }
-      
-      if (!isRoadTile(currentGrid, currentGridSize, car.tileX, car.tileY)) {
-        continue;
-      }
-      
-      car.progress += car.speed * delta * speedMultiplier;
-      let guard = 0;
-      while (car.progress >= 1 && guard < 4) {
-        guard++;
-        const meta = DIRECTION_META[car.direction];
-        car.tileX += meta.step.x;
-        car.tileY += meta.step.y;
-        
-        if (!isRoadTile(currentGrid, currentGridSize, car.tileX, car.tileY)) {
-          alive = false;
-          break;
-        }
-        
-        car.progress -= 1;
-        const nextDirection = pickNextDirection(car.direction, currentGrid, currentGridSize, car.tileX, car.tileY);
-        if (!nextDirection) {
-          alive = false;
-          break;
-        }
-        car.direction = nextDirection;
-      }
-      
-      if (alive) {
-        updatedCars.push(car);
-      }
-    }
-    
-    carsRef.current = updatedCars;
-  }, [spawnRandomCar, isMobile]);
-
-  // Update pedestrians - only when zoomed in enough
-  const updatePedestrians = useCallback((delta: number) => {
-    const { grid: currentGrid, gridSize: currentGridSize, speed: currentSpeed, zoom: currentZoom } = worldStateRef.current;
-    
-    // Clear pedestrians if zoomed out (mobile requires higher zoom level)
-    const minZoomForPedestrians = isMobile ? 0.8 : PEDESTRIAN_MIN_ZOOM;
-    if (currentZoom < minZoomForPedestrians) {
-      pedestriansRef.current = [];
-      return;
-    }
-    
-    if (!currentGrid || currentGridSize <= 0) {
-      pedestriansRef.current = [];
-      return;
-    }
-    
-    // Speed multiplier
-    const speedMultiplier = currentSpeed === 0 ? 0 : currentSpeed === 1 ? 1 : currentSpeed === 2 ? 2.5 : 4;
-    
-    // Get cached road tile count (only recalculate when grid changes)
-    const currentGridVersion = gridVersionRef.current;
-    let roadTileCount: number;
-    if (cachedRoadTileCountRef.current.gridVersion === currentGridVersion) {
-      roadTileCount = cachedRoadTileCountRef.current.count;
-    } else {
-      // Recalculate and cache
-      roadTileCount = 0;
-      for (let y = 0; y < currentGridSize; y++) {
-        for (let x = 0; x < currentGridSize; x++) {
-          if (currentGrid[y][x].building.type === 'road') {
-            roadTileCount++;
-          }
-        }
-      }
-      cachedRoadTileCountRef.current = { count: roadTileCount, gridVersion: currentGridVersion };
-    }
-    
-    // Spawn pedestrians - scale with road network size, reduced on mobile
-    // Mobile: max 50 pedestrians, 0.8 per road tile
-    // Desktop: max 200+ pedestrians, 3 per road tile
-    const maxPedestrians = isMobile 
-      ? Math.min(50, Math.max(20, Math.floor(roadTileCount * 0.8)))
-      : Math.max(200, roadTileCount * 3);
-    pedestrianSpawnTimerRef.current -= delta;
-    if (pedestriansRef.current.length < maxPedestrians && pedestrianSpawnTimerRef.current <= 0) {
-      // Spawn fewer pedestrians at once on mobile
-      let spawnedCount = 0;
-      const spawnBatch = isMobile 
-        ? Math.min(8, Math.max(3, Math.floor(roadTileCount / 25)))
-        : Math.min(50, Math.max(20, Math.floor(roadTileCount / 10)));
-      for (let i = 0; i < spawnBatch; i++) {
-        if (spawnPedestrian()) {
-          spawnedCount++;
-        }
-      }
-      // Slower spawn rate on mobile
-      pedestrianSpawnTimerRef.current = spawnedCount > 0 ? (isMobile ? 0.15 : 0.02) : (isMobile ? 0.08 : 0.01);
-    }
-    
-    const updatedPedestrians: Pedestrian[] = [];
-    
-    for (const ped of [...pedestriansRef.current]) {
-      let alive = true;
-      
-      // Update age
-      ped.age += delta;
-      if (ped.age > ped.maxAge) {
-        continue;
-      }
-      
-      // Update walk animation
-      ped.walkOffset += delta * 8;
-      
-      // Check if still on valid road
-      if (!isRoadTile(currentGrid, currentGridSize, ped.tileX, ped.tileY)) {
-        continue;
-      }
-      
-      // Move pedestrian along path
-      ped.progress += ped.speed * delta * speedMultiplier;
-      
-      // Handle single-tile paths (already at destination)
-      if (ped.path.length === 1 && ped.progress >= 1) {
-        if (!ped.returningHome) {
-          ped.returningHome = true;
-          const returnPath = findPathOnRoads(currentGrid, currentGridSize, ped.destX, ped.destY, ped.homeX, ped.homeY);
-          if (returnPath && returnPath.length > 0) {
-            ped.path = returnPath;
-            ped.pathIndex = 0;
-            ped.progress = 0;
-            ped.tileX = returnPath[0].x;
-            ped.tileY = returnPath[0].y;
-            if (returnPath.length > 1) {
-              const nextTile = returnPath[1];
-              const dir = getDirectionToTile(returnPath[0].x, returnPath[0].y, nextTile.x, nextTile.y);
-              if (dir) ped.direction = dir;
-            }
-          } else {
-            continue; // Remove pedestrian
-          }
-        } else {
-          continue; // Arrived home, remove
-        }
-      }
-      
-      while (ped.progress >= 1 && ped.pathIndex < ped.path.length - 1) {
-        ped.pathIndex++;
-        ped.progress -= 1;
-        
-        const currentTile = ped.path[ped.pathIndex];
-        
-        // Bounds check
-        if (currentTile.x < 0 || currentTile.x >= currentGridSize ||
-            currentTile.y < 0 || currentTile.y >= currentGridSize) {
-          alive = false;
-          break;
-        }
-        
-        ped.tileX = currentTile.x;
-        ped.tileY = currentTile.y;
-        
-        // Check if reached end of path
-        if (ped.pathIndex >= ped.path.length - 1) {
-          if (!ped.returningHome) {
-            // Arrived at destination - start returning home
-            ped.returningHome = true;
-            const returnPath = findPathOnRoads(currentGrid, currentGridSize, ped.destX, ped.destY, ped.homeX, ped.homeY);
-            if (returnPath && returnPath.length > 0) {
-              ped.path = returnPath;
-              ped.pathIndex = 0;
-              ped.progress = 0;
-              // Update direction for return trip
-              if (returnPath.length > 1) {
-                const nextTile = returnPath[1];
-                const dir = getDirectionToTile(returnPath[0].x, returnPath[0].y, nextTile.x, nextTile.y);
-                if (dir) ped.direction = dir;
-              }
-            } else {
-              alive = false;
-            }
-          } else {
-            // Arrived back home - remove pedestrian
-            alive = false;
-          }
-          break;
-        }
-        
-        // Update direction for next segment
-        if (ped.pathIndex + 1 < ped.path.length) {
-          const nextTile = ped.path[ped.pathIndex + 1];
-          const dir = getDirectionToTile(ped.tileX, ped.tileY, nextTile.x, nextTile.y);
-          if (dir) ped.direction = dir;
-        }
-      }
-      
-      // Handle case where pedestrian is already at the last tile with progress >= 1
-      // (can happen when spawned at end of path, or if progress accumulates)
-      if (alive && ped.progress >= 1 && ped.pathIndex >= ped.path.length - 1) {
-        if (!ped.returningHome) {
-          // Arrived at destination - start returning home
-          ped.returningHome = true;
-          const returnPath = findPathOnRoads(currentGrid, currentGridSize, ped.destX, ped.destY, ped.homeX, ped.homeY);
-          if (returnPath && returnPath.length > 0) {
-            ped.path = returnPath;
-            ped.pathIndex = 0;
-            ped.progress = 0;
-            ped.tileX = returnPath[0].x;
-            ped.tileY = returnPath[0].y;
-            // Update direction for return trip
-            if (returnPath.length > 1) {
-              const nextTile = returnPath[1];
-              const dir = getDirectionToTile(returnPath[0].x, returnPath[0].y, nextTile.x, nextTile.y);
-              if (dir) ped.direction = dir;
-            }
-          } else {
-            alive = false;
-          }
-        } else {
-          // Arrived back home - remove pedestrian
-          alive = false;
-        }
-      }
-      
-      if (alive) {
-        updatedPedestrians.push(ped);
-      }
-    }
-    
-    pedestriansRef.current = updatedPedestrians;
-  }, [spawnPedestrian, isMobile]);
+  // Pedestrian updating now handled by usePedestrians hook
 
   const drawCars = useCallback((ctx: CanvasRenderingContext2D) => {
-    const { offset: currentOffset, zoom: currentZoom, grid: currentGrid, gridSize: currentGridSize } = worldStateRef.current;
-    const canvas = ctx.canvas;
-    const dpr = window.devicePixelRatio || 1;
-    
-    ctx.setTransform(1, 0, 0, 1, 0, 0);
-    ctx.clearRect(0, 0, canvas.width, canvas.height);
-    
-    // Early exit if no grid data
-    if (!currentGrid || currentGridSize <= 0 || carsRef.current.length === 0) {
-      return;
-    }
-    
-    ctx.save();
-    ctx.scale(dpr * currentZoom, dpr * currentZoom);
-    ctx.translate(currentOffset.x / currentZoom, currentOffset.y / currentZoom);
-    
-    const viewWidth = canvas.width / (dpr * currentZoom);
-    const viewHeight = canvas.height / (dpr * currentZoom);
-    const viewLeft = -currentOffset.x / currentZoom - TILE_WIDTH;
-    const viewTop = -currentOffset.y / currentZoom - TILE_HEIGHT * 2;
-    const viewRight = viewWidth - currentOffset.x / currentZoom + TILE_WIDTH;
-    const viewBottom = viewHeight - currentOffset.y / currentZoom + TILE_HEIGHT * 2;
-    
-    // Helper function to check if a car is behind a building
-    const isCarBehindBuilding = (carTileX: number, carTileY: number): boolean => {
-      // Only check tiles directly in front (higher depth means drawn later/on top)
-      const carDepth = carTileX + carTileY;
-      
-      // Check a small area - just tiles that could visually cover the car
-      for (let dy = 0; dy <= 1; dy++) {
-        for (let dx = 0; dx <= 1; dx++) {
-          if (dx === 0 && dy === 0) continue; // Skip the car's own tile
-          
-          const checkX = carTileX + dx;
-          const checkY = carTileY + dy;
-          
-          // Skip if out of bounds
-          if (checkX < 0 || checkY < 0 || checkX >= currentGridSize || checkY >= currentGridSize) {
-            continue;
-          }
-          
-          const tile = currentGrid[checkY]?.[checkX];
-          if (!tile) continue;
-          
-          const buildingType = tile.building.type;
-          
-          // Skip roads, grass, empty, water, and trees (these don't hide cars)
-          const skipTypes: BuildingType[] = ['road', 'grass', 'empty', 'water', 'tree'];
-          if (skipTypes.includes(buildingType)) {
-            continue;
-          }
-          
-          // Check if this building tile has higher depth (drawn after/on top)
-          const buildingDepth = checkX + checkY;
-          
-          // Only hide if building is strictly in front (higher depth)
-          if (buildingDepth > carDepth) {
-            return true;
-          }
-        }
-      }
-      
-      return false;
-    };
-    
-    carsRef.current.forEach(car => {
-      const { screenX, screenY } = gridToScreen(car.tileX, car.tileY, 0, 0);
-      const centerX = screenX + TILE_WIDTH / 2;
-      const centerY = screenY + TILE_HEIGHT / 2;
-      const meta = DIRECTION_META[car.direction];
-      const carX = centerX + meta.vec.dx * car.progress + meta.normal.nx * car.laneOffset;
-      const carY = centerY + meta.vec.dy * car.progress + meta.normal.ny * car.laneOffset;
-      
-      if (carX < viewLeft - 40 || carX > viewRight + 40 || carY < viewTop - 60 || carY > viewBottom + 60) {
-        return;
-      }
-      
-      // Check if car is behind a building - if so, skip drawing
-      if (isCarBehindBuilding(car.tileX, car.tileY)) {
-        return;
-      }
-      
-      ctx.save();
-      ctx.translate(carX, carY);
-      ctx.rotate(meta.angle);
-      
-      // Scale down by 30% (multiply by 0.7)
-      const scale = 0.7;
-      
-      ctx.fillStyle = car.color;
-      ctx.beginPath();
-      ctx.moveTo(-10 * scale, -5 * scale);
-      ctx.lineTo(10 * scale, -5 * scale);
-      ctx.lineTo(12 * scale, 0);
-      ctx.lineTo(10 * scale, 5 * scale);
-      ctx.lineTo(-10 * scale, 5 * scale);
-      ctx.closePath();
-      ctx.fill();
-      
-      // Windshield
-      ctx.fillStyle = 'rgba(255, 255, 255, 0.6)';
-      ctx.fillRect(-4 * scale, -2.8 * scale, 7 * scale, 5.6 * scale);
-      
-      // Rear
-      ctx.fillStyle = '#111827';
-      ctx.fillRect(-10 * scale, -4 * scale, 2.4 * scale, 8 * scale);
-      
-      ctx.restore();
-    });
-    
-    ctx.restore();
-  }, []);
+    drawCarsUtil(ctx, carsRef.current, worldStateRef.current, grid, gridSize);
+  }, [grid, gridSize]);
 
   // Draw pedestrians with simple SVG-style sprites (uses extracted utility)
   const drawPedestrians = useCallback((ctx: CanvasRenderingContext2D) => {
@@ -2771,148 +1996,8 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
 
   // Draw emergency vehicles (fire trucks and police cars)
   const drawEmergencyVehicles = useCallback((ctx: CanvasRenderingContext2D) => {
-    const { offset: currentOffset, zoom: currentZoom, grid: currentGrid, gridSize: currentGridSize } = worldStateRef.current;
-    const canvas = ctx.canvas;
-    const dpr = window.devicePixelRatio || 1;
-    
-    // Early exit if no emergency vehicles
-    if (!currentGrid || currentGridSize <= 0 || emergencyVehiclesRef.current.length === 0) {
-      return;
-    }
-    
-    ctx.save();
-    ctx.scale(dpr * currentZoom, dpr * currentZoom);
-    ctx.translate(currentOffset.x / currentZoom, currentOffset.y / currentZoom);
-    
-    const viewWidth = canvas.width / (dpr * currentZoom);
-    const viewHeight = canvas.height / (dpr * currentZoom);
-    const viewLeft = -currentOffset.x / currentZoom - TILE_WIDTH;
-    const viewTop = -currentOffset.y / currentZoom - TILE_HEIGHT * 2;
-    const viewRight = viewWidth - currentOffset.x / currentZoom + TILE_WIDTH;
-    const viewBottom = viewHeight - currentOffset.y / currentZoom + TILE_HEIGHT * 2;
-    
-    // Helper function to check if a vehicle is behind a building
-    const isVehicleBehindBuilding = (tileX: number, tileY: number): boolean => {
-      const vehicleDepth = tileX + tileY;
-      
-      for (let dy = 0; dy <= 1; dy++) {
-        for (let dx = 0; dx <= 1; dx++) {
-          if (dx === 0 && dy === 0) continue;
-          
-          const checkX = tileX + dx;
-          const checkY = tileY + dy;
-          
-          if (checkX < 0 || checkY < 0 || checkX >= currentGridSize || checkY >= currentGridSize) {
-            continue;
-          }
-          
-          const tile = currentGrid[checkY]?.[checkX];
-          if (!tile) continue;
-          
-          const buildingType = tile.building.type;
-          const skipTypes: BuildingType[] = ['road', 'grass', 'empty', 'water', 'tree'];
-          if (skipTypes.includes(buildingType)) {
-            continue;
-          }
-          
-          const buildingDepth = checkX + checkY;
-          if (buildingDepth > vehicleDepth) {
-            return true;
-          }
-        }
-      }
-      
-      return false;
-    };
-    
-    emergencyVehiclesRef.current.forEach(vehicle => {
-      const { screenX, screenY } = gridToScreen(vehicle.tileX, vehicle.tileY, 0, 0);
-      const centerX = screenX + TILE_WIDTH / 2;
-      const centerY = screenY + TILE_HEIGHT / 2;
-      const meta = DIRECTION_META[vehicle.direction];
-      const vehicleX = centerX + meta.vec.dx * vehicle.progress + meta.normal.nx * vehicle.laneOffset;
-      const vehicleY = centerY + meta.vec.dy * vehicle.progress + meta.normal.ny * vehicle.laneOffset;
-      
-      // View culling
-      if (vehicleX < viewLeft - 40 || vehicleX > viewRight + 40 || vehicleY < viewTop - 60 || vehicleY > viewBottom + 60) {
-        return;
-      }
-      
-      ctx.save();
-      ctx.translate(vehicleX, vehicleY);
-      ctx.rotate(meta.angle);
-      
-      const scale = 0.6; // Smaller emergency vehicles
-      
-      // Vehicle body color
-      const bodyColor = vehicle.type === 'fire_truck' ? '#dc2626' : '#1e40af';
-      
-      // Draw vehicle body (longer for fire trucks)
-      const length = vehicle.type === 'fire_truck' ? 14 : 11;
-      ctx.fillStyle = bodyColor;
-      ctx.beginPath();
-      ctx.moveTo(-length * scale, -5 * scale);
-      ctx.lineTo(length * scale, -5 * scale);
-      ctx.lineTo((length + 2) * scale, 0);
-      ctx.lineTo(length * scale, 5 * scale);
-      ctx.lineTo(-length * scale, 5 * scale);
-      ctx.closePath();
-      ctx.fill();
-      
-      // Draw stripe/accent
-      ctx.fillStyle = vehicle.type === 'fire_truck' ? '#fbbf24' : '#ffffff';
-      ctx.fillRect(-length * scale * 0.5, -3 * scale, length * scale, 6 * scale * 0.3);
-      
-      // Draw windshield
-      ctx.fillStyle = 'rgba(200, 220, 255, 0.7)';
-      ctx.fillRect(-2 * scale, -3 * scale, 5 * scale, 6 * scale);
-      
-      // Draw emergency lights (flashing)
-      const flashOn = Math.sin(vehicle.flashTimer) > 0;
-      const flashOn2 = Math.sin(vehicle.flashTimer + Math.PI) > 0;
-      
-      // Light bar on top
-      if (vehicle.type === 'fire_truck') {
-        // Fire truck has red lights
-        ctx.fillStyle = flashOn ? '#ff0000' : '#880000';
-        ctx.fillRect(-6 * scale, -7 * scale, 3 * scale, 3 * scale);
-        ctx.fillStyle = flashOn2 ? '#ff0000' : '#880000';
-        ctx.fillRect(3 * scale, -7 * scale, 3 * scale, 3 * scale);
-        
-        // Glow effect
-        if (flashOn || flashOn2) {
-          ctx.shadowColor = '#ff0000';
-          ctx.shadowBlur = 6;
-          ctx.fillStyle = 'rgba(255, 0, 0, 0.4)';
-          ctx.fillRect(-8 * scale, -8 * scale, 16 * scale, 4 * scale);
-          ctx.shadowBlur = 0;
-        }
-      } else {
-        // Police car has red and blue lights
-        ctx.fillStyle = flashOn ? '#ff0000' : '#880000';
-        ctx.fillRect(-5 * scale, -7 * scale, 3 * scale, 3 * scale);
-        ctx.fillStyle = flashOn2 ? '#0066ff' : '#003388';
-        ctx.fillRect(2 * scale, -7 * scale, 3 * scale, 3 * scale);
-        
-        // Glow effect
-        if (flashOn || flashOn2) {
-          ctx.shadowColor = flashOn ? '#ff0000' : '#0066ff';
-          ctx.shadowBlur = 6;
-          ctx.fillStyle = flashOn ? 'rgba(255, 0, 0, 0.4)' : 'rgba(0, 100, 255, 0.4)';
-          ctx.fillRect(-7 * scale, -8 * scale, 14 * scale, 4 * scale);
-          ctx.shadowBlur = 0;
-        }
-      }
-      
-      // Draw rear wheels/details
-      ctx.fillStyle = '#111827';
-      ctx.fillRect(-length * scale, -4 * scale, 2 * scale, 8 * scale);
-      
-      ctx.restore();
-    });
-    
-    ctx.restore();
-  }, []);
+    drawEmergencyVehiclesUtil(ctx, emergencyVehiclesRef.current, worldStateRef.current, grid, gridSize);
+  }, [grid, gridSize]);
 
   // Animation time ref for incident indicator pulsing
   const incidentAnimTimeRef = useRef(0);
@@ -3115,124 +2200,11 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
   }, [currentSpritePack]);
   
   // Helper function to check if a tile is part of a multi-tile building footprint
-  const isPartOfMultiTileBuilding = useCallback((gridX: number, gridY: number): boolean => {
-    // Check all possible origin positions that could have a multi-tile building covering this tile
-    // For a 2x2 building, check up to 1 tile away in each direction
-    // For a 3x3 building, check up to 2 tiles away
-    // For a 4x4 building, check up to 3 tiles away
-    const maxSize = 4; // Maximum building size
-    
-    for (let dy = 0; dy < maxSize; dy++) {
-      for (let dx = 0; dx < maxSize; dx++) {
-        const originX = gridX - dx;
-        const originY = gridY - dy;
-        
-        if (originX >= 0 && originX < gridSize && originY >= 0 && originY < gridSize) {
-          const originTile = grid[originY][originX];
-          const buildingSize = getBuildingSize(originTile.building.type);
-          
-          // Check if this tile is within the footprint of the building at origin
-          if (buildingSize.width > 1 || buildingSize.height > 1) {
-            if (gridX >= originX && gridX < originX + buildingSize.width &&
-                gridY >= originY && gridY < originY + buildingSize.height) {
-              return true;
-            }
-          }
-        }
-      }
-    }
-    
-    return false;
+  // Building helper functions now extracted to buildingHelpers.ts
+  // Memoize findBuildingOrigin for use in event handlers
+  const findBuildingOriginMemo = useMemo(() => {
+    return (gridX: number, gridY: number) => findBuildingOrigin(gridX, gridY, grid, gridSize);
   }, [grid, gridSize]);
-  
-  // Helper function to find the origin of a multi-tile building that contains a given tile
-  // Returns the origin coordinates and building type, or null if not part of a multi-tile building
-  const findBuildingOrigin = useCallback((gridX: number, gridY: number): { originX: number; originY: number; buildingType: BuildingType } | null => {
-    const maxSize = 4; // Maximum building size
-    
-    // First check if this tile itself has a multi-tile building
-    const tile = grid[gridY]?.[gridX];
-    if (!tile) return null;
-    
-    // If this tile has a real building (not empty), check if it's multi-tile
-    if (tile.building.type !== 'empty' && 
-        tile.building.type !== 'grass' && 
-        tile.building.type !== 'water' && 
-        tile.building.type !== 'road' && 
-        tile.building.type !== 'tree') {
-      const size = getBuildingSize(tile.building.type);
-      if (size.width > 1 || size.height > 1) {
-        return { originX: gridX, originY: gridY, buildingType: tile.building.type };
-      }
-      return null; // Single-tile building
-    }
-    
-    // If this is an 'empty' tile, search for the origin building
-    if (tile.building.type === 'empty') {
-      for (let dy = 0; dy < maxSize; dy++) {
-        for (let dx = 0; dx < maxSize; dx++) {
-          const originX = gridX - dx;
-          const originY = gridY - dy;
-          
-          if (originX >= 0 && originX < gridSize && originY >= 0 && originY < gridSize) {
-            const originTile = grid[originY][originX];
-            
-            if (originTile.building.type !== 'empty' && 
-                originTile.building.type !== 'grass' &&
-                originTile.building.type !== 'water' &&
-                originTile.building.type !== 'road' &&
-                originTile.building.type !== 'tree') {
-              const size = getBuildingSize(originTile.building.type);
-              
-              // Check if the clicked tile is within this building's footprint
-              if (size.width > 1 || size.height > 1) {
-                if (gridX >= originX && gridX < originX + size.width &&
-                    gridY >= originY && gridY < originY + size.height) {
-                  return { originX, originY, buildingType: originTile.building.type };
-                }
-              }
-            }
-          }
-        }
-      }
-    }
-    
-    return null;
-  }, [grid, gridSize]);
-  
-// PERF: Static Set for park building lookups (O(1) instead of O(n) array includes)
-  const parkBuildingsSet = useMemo(() => new Set<BuildingType>([
-    'park_large', 'baseball_field_small', 'football_field',
-    'mini_golf_course', 'go_kart_track', 'amphitheater', 'greenhouse_garden',
-    'marina_docks_small', 'roller_coaster_small', 'mountain_lodge', 'playground_large', 'mountain_trailhead'
-  ]), []);
-  
-  // Helper function to check if a tile is part of a park building footprint
-  // Note: buildings with grey bases (baseball_stadium, swimming_pool, community_center, office_building_small) are NOT included
-  const isPartOfParkBuilding = useCallback((gridX: number, gridY: number): boolean => {
-    const maxSize = 4; // Maximum building size
-
-    for (let dy = 0; dy < maxSize; dy++) {
-      for (let dx = 0; dx < maxSize; dx++) {
-        const originX = gridX - dx;
-        const originY = gridY - dy;
-
-        if (originX >= 0 && originX < gridSize && originY >= 0 && originY < gridSize) {
-          const originTile = grid[originY][originX];
-
-          // PERF: Use Set.has() instead of array.includes() for O(1) lookup
-          if (parkBuildingsSet.has(originTile.building.type)) {
-            const buildingSize = getBuildingSize(originTile.building.type);
-            if (gridX >= originX && gridX < originX + buildingSize.width &&
-                gridY >= originY && gridY < originY + buildingSize.height) {
-              return true;
-            }
-          }
-        }
-      }
-    }
-    return false;
-  }, [grid, gridSize, parkBuildingsSet]);
   
   // Update canvas size on resize with high-DPI support
   useEffect(() => {
@@ -3799,7 +2771,7 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
         'pier_large', 'roller_coaster_small', 'community_garden', 'pond_park', 'park_gate', 
         'mountain_lodge', 'mountain_trailhead'];
       const isPark = allParkTypes.includes(tile.building.type) ||
-                     (tile.building.type === 'empty' && isPartOfParkBuilding(tile.x, tile.y));
+                     (tile.building.type === 'empty' && isPartOfParkBuilding(tile.x, tile.y, grid, gridSize));
       // Check if this is a building (not grass, empty, water, road, tree, or parks)
       // Also check if it's part of a multi-tile building footprint
       const isDirectBuilding = !isPark &&
@@ -3808,7 +2780,7 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
         tile.building.type !== 'water' &&
         tile.building.type !== 'road' &&
         tile.building.type !== 'tree';
-      const isPartOfBuilding = tile.building.type === 'empty' && isPartOfMultiTileBuilding(tile.x, tile.y);
+      const isPartOfBuilding = tile.building.type === 'empty' && isPartOfMultiTileBuilding(tile.x, tile.y, grid, gridSize);
       const isBuilding = isDirectBuilding || isPartOfBuilding;
       
       // ALL buildings get grey/concrete base tiles (except parks which stay green)
@@ -4531,14 +3503,14 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
           'pier_large', 'roller_coaster_small', 'community_garden', 'pond_park', 'park_gate', 
           'mountain_lodge', 'mountain_trailhead'];
         const isPark = allParkTypesCheck.includes(tile.building.type) ||
-                       (tile.building.type === 'empty' && isPartOfParkBuilding(x, y));
+                       (tile.building.type === 'empty' && isPartOfParkBuilding(x, y, grid, gridSize));
         const isDirectBuilding = !isPark &&
           tile.building.type !== 'grass' &&
           tile.building.type !== 'empty' &&
           tile.building.type !== 'water' &&
           tile.building.type !== 'road' &&
           tile.building.type !== 'tree';
-        const isPartOfBuilding = tile.building.type === 'empty' && isPartOfMultiTileBuilding(x, y);
+        const isPartOfBuilding = tile.building.type === 'empty' && isPartOfMultiTileBuilding(x, y, grid, gridSize);
         const needsGreyBase = (isDirectBuilding || isPartOfBuilding) && !isPark;
         
         // Check if this is a grass/empty tile adjacent to water (needs green base drawn over water)
@@ -4547,7 +3519,7 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
         
         // Check if this is a park that needs a green base tile
         const needsGreenBaseForPark = (tile.building.type === 'park' || tile.building.type === 'park_large') ||
-                                      (tile.building.type === 'empty' && isPartOfParkBuilding(x, y));
+                                      (tile.building.type === 'empty' && isPartOfParkBuilding(x, y, grid, gridSize));
         
         // Draw base tile for all tiles (including water), but skip gray bases for buildings and green bases for grass/empty adjacent to water or parks
         // Highlight subway stations when subway overlay is active
@@ -4785,7 +3757,7 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
       if (delta > 0) {
         updateCars(delta);
         spawnCrimeIncidents(delta); // Spawn new crime incidents
-        updateCrimeIncidents(delta); // Update/decay crime incidents
+        updateCrimeIncidents(delta, activeCrimesRef); // Update/decay crime incidents
         updateEmergencyVehicles(delta); // Update emergency vehicles!
         updatePedestrians(delta); // Update pedestrians (zoom-gated)
         updateAirplanes(delta); // Update airplanes (airport required)
@@ -5091,7 +4063,7 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
         if (gridX >= 0 && gridX < gridSize && gridY >= 0 && gridY < gridSize) {
           if (selectedTool === 'select') {
             // For multi-tile buildings, select the origin tile
-            const origin = findBuildingOrigin(gridX, gridY);
+            const origin = findBuildingOriginMemo(gridX, gridY);
             if (origin) {
               setSelectedTile({ x: origin.originX, y: origin.originY });
             } else {
@@ -5120,35 +4092,13 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
         }
       }
     }
-  }, [offset, gridSize, selectedTool, placeAtTile, zoom, showsDragGrid, supportsDragPlace, setSelectedTile, findBuildingOrigin]);
+  }, [offset, gridSize, selectedTool, placeAtTile, zoom, showsDragGrid, supportsDragPlace, setSelectedTile, findBuildingOriginMemo]);
   
   // Calculate camera bounds based on grid size
-  const getMapBounds = useCallback((currentZoom: number, canvasW: number, canvasH: number) => {
-    const n = gridSize;
-    const padding = 100; // Allow some over-scroll
-    
-    // Map bounds in world coordinates
-    const mapLeft = -(n - 1) * TILE_WIDTH / 2;
-    const mapRight = (n - 1) * TILE_WIDTH / 2;
-    const mapTop = 0;
-    const mapBottom = (n - 1) * TILE_HEIGHT;
-    
-    const minOffsetX = padding - mapRight * currentZoom;
-    const maxOffsetX = canvasW - padding - mapLeft * currentZoom;
-    const minOffsetY = padding - mapBottom * currentZoom;
-    const maxOffsetY = canvasH - padding - mapTop * currentZoom;
-    
-    return { minOffsetX, maxOffsetX, minOffsetY, maxOffsetY };
+  // Viewport helper functions now extracted to viewportHelpers.ts
+  const getMapBounds = useMemo(() => (currentZoom: number, canvasW: number, canvasH: number) => {
+    return getMapBoundsUtil(currentZoom, canvasW, canvasH, gridSize);
   }, [gridSize]);
-  
-  // Clamp offset to keep camera within reasonable bounds
-  const clampOffset = useCallback((newOffset: { x: number; y: number }, currentZoom: number) => {
-    const bounds = getMapBounds(currentZoom, canvasSize.width, canvasSize.height);
-    return {
-      x: Math.max(bounds.minOffsetX, Math.min(bounds.maxOffsetX, newOffset.x)),
-      y: Math.max(bounds.minOffsetY, Math.min(bounds.maxOffsetY, newOffset.y)),
-    };
-  }, [getMapBounds, canvasSize.width, canvasSize.height]);
 
   // Handle minimap navigation - center the view on the target tile
   useEffect(() => {
@@ -5183,7 +4133,7 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
         x: e.clientX - dragStart.x,
         y: e.clientY - dragStart.y,
       };
-      setOffset(clampOffset(newOffset, zoom));
+      setOffset(clampOffset(newOffset, zoom, canvasSize.width, canvasSize.height, gridSize));
       return;
     }
     
@@ -5352,19 +4302,7 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
     setZoom(newZoom);
   }, [zoom, offset, clampOffset]);
 
-  // Touch handlers for mobile
-  const getTouchDistance = useCallback((touch1: React.Touch, touch2: React.Touch) => {
-    const dx = touch1.clientX - touch2.clientX;
-    const dy = touch1.clientY - touch2.clientY;
-    return Math.sqrt(dx * dx + dy * dy);
-  }, []);
-
-  const getTouchCenter = useCallback((touch1: React.Touch, touch2: React.Touch) => {
-    return {
-      x: (touch1.clientX + touch2.clientX) / 2,
-      y: (touch1.clientY + touch2.clientY) / 2,
-    };
-  }, []);
+  // Touch helper functions now extracted to viewportHelpers.ts
 
   const handleTouchStart = useCallback((e: React.TouchEvent) => {
     if (e.touches.length === 1) {
@@ -5393,7 +4331,7 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
         x: touch.clientX - dragStart.x,
         y: touch.clientY - dragStart.y,
       };
-      setOffset(clampOffset(newOffset, zoom));
+      setOffset(clampOffset(newOffset, zoom, canvasSize.width, canvasSize.height, gridSize));
     } else if (e.touches.length === 2 && initialPinchDistanceRef.current !== null) {
       // Pinch to zoom
       const currentDistance = getTouchDistance(e.touches[0], e.touches[1]);
@@ -5422,7 +4360,10 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
 
         const clampedOffset = clampOffset(
           { x: newOffsetX + panDeltaX, y: newOffsetY + panDeltaY },
-          newZoom
+          newZoom,
+          canvasSize.width,
+          canvasSize.height,
+          gridSize
         );
 
         setOffset(clampedOffset);
@@ -5453,7 +4394,7 @@ export function CanvasIsometricGrid({ overlayMode, selectedTile, setSelectedTile
 
             if (gridX >= 0 && gridX < gridSize && gridY >= 0 && gridY < gridSize) {
               if (selectedTool === 'select') {
-                const origin = findBuildingOrigin(gridX, gridY);
+                const origin = findBuildingOriginMemo(gridX, gridY);
                 if (origin) {
                   setSelectedTile({ x: origin.originX, y: origin.originY });
                 } else {
