+++ src/components/game/drawing/index.ts
@@ -1,0 +1,2 @@
+export { drawCars } from './drawCars';
+export { drawEmergencyVehicles } from './drawEmergencyVehicles';
