+++ src/components/game/hooks/index.ts
@@ -1,0 +1,4 @@
+export { useCars } from './useCars';
+export { useCrimeIncidents, type CrimeIncident } from './useCrimeIncidents';
+export { usePedestrians } from './usePedestrians';
+export { useEmergencyVehicles } from './useEmergencyVehicles';
