+++ src/components/game/hooks/useCars.ts
@@ -1,0 +1,110 @@
+import { useRef, useCallback } from 'react';
+import { Car, WorldRenderState, CAR_COLORS, DIRECTION_META } from '@/components/game/types';
+import { isRoadTile, getDirectionOptions, pickNextDirection } from '@/components/game/utils';
+
+export function useCars(worldStateRef: React.MutableRefObject<WorldRenderState>, isMobile: boolean) {
+  const carsRef = useRef<Car[]>([]);
+  const carIdRef = useRef(0);
+  const carSpawnTimerRef = useRef(0);
+
+  const spawnRandomCar = useCallback(() => {
+    const { grid: currentGrid, gridSize: currentGridSize } = worldStateRef.current;
+    if (!currentGrid || currentGridSize <= 0) return false;
+    
+    for (let attempt = 0; attempt < 20; attempt++) {
+      const tileX = Math.floor(Math.random() * currentGridSize);
+      const tileY = Math.floor(Math.random() * currentGridSize);
+      if (!isRoadTile(currentGrid, currentGridSize, tileX, tileY)) continue;
+      
+      const options = getDirectionOptions(currentGrid, currentGridSize, tileX, tileY);
+      if (options.length === 0) continue;
+      
+      const direction = options[Math.floor(Math.random() * options.length)];
+      carsRef.current.push({
+        id: carIdRef.current++,
+        tileX,
+        tileY,
+        direction,
+        progress: Math.random() * 0.8,
+        speed: (0.35 + Math.random() * 0.35) * 0.7,
+        age: 0,
+        maxAge: 1800 + Math.random() * 2700,
+        color: CAR_COLORS[Math.floor(Math.random() * CAR_COLORS.length)],
+        laneOffset: (Math.random() < 0.5 ? -1 : 1) * (4 + Math.random() * 3),
+      });
+      return true;
+    }
+    
+    return false;
+  }, [worldStateRef]);
+
+  const updateCars = useCallback((delta: number) => {
+    const { grid: currentGrid, gridSize: currentGridSize, speed: currentSpeed } = worldStateRef.current;
+    if (!currentGrid || currentGridSize <= 0) {
+      carsRef.current = [];
+      return;
+    }
+    
+    // Speed multiplier: 0 = paused, 1 = normal, 2 = fast (2x), 3 = very fast (4x)
+    const speedMultiplier = currentSpeed === 0 ? 0 : currentSpeed === 1 ? 1 : currentSpeed === 2 ? 2.5 : 4;
+    
+    // Reduce max cars on mobile for better performance
+    const baseMaxCars = 160;
+    const maxCars = Math.min(baseMaxCars, Math.max(16, Math.floor(currentGridSize * (2))));
+    carSpawnTimerRef.current -= delta;
+    if (carsRef.current.length < maxCars && carSpawnTimerRef.current <= 0) {
+      if (spawnRandomCar()) {
+        carSpawnTimerRef.current = 0.9 + Math.random() * 1.3;
+      } else {
+        carSpawnTimerRef.current = 0.5;
+      }
+    }
+    
+    const updatedCars: Car[] = [];
+    for (const car of [...carsRef.current]) {
+      let alive = true;
+      
+      car.age += delta;
+      if (car.age > car.maxAge) {
+        continue;
+      }
+      
+      if (!isRoadTile(currentGrid, currentGridSize, car.tileX, car.tileY)) {
+        continue;
+      }
+      
+      car.progress += car.speed * delta * speedMultiplier;
+      let guard = 0;
+      while (car.progress >= 1 && guard < 4) {
+        guard++;
+        const meta = DIRECTION_META[car.direction];
+        car.tileX += meta.step.x;
+        car.tileY += meta.step.y;
+        
+        if (!isRoadTile(currentGrid, currentGridSize, car.tileX, car.tileY)) {
+          alive = false;
+          break;
+        }
+        
+        car.progress -= 1;
+        const nextDirection = pickNextDirection(car.direction, currentGrid, currentGridSize, car.tileX, car.tileY);
+        if (!nextDirection) {
+          alive = false;
+          break;
+        }
+        car.direction = nextDirection;
+      }
+      
+      if (alive) {
+        updatedCars.push(car);
+      }
+    }
+    
+    carsRef.current = updatedCars;
+  }, [spawnRandomCar, worldStateRef]);
+
+  return {
+    carsRef,
+    updateCars,
+  };
+}
